console: generalize async ExecutionRegistry into the API
Move ExecutionRegistry to the api module and make it accept any ScriptContext
(plus an optional ExecutionCallback run on the worker thread), so extensions can
run scripts asynchronously with their own context. AsyncScriptContext moves to
api context.impl to avoid a split package now that the streaming package is
exported.
