+++ api/src/main/groovy/be/orbinson/aem/groovy/console/streaming/package-info.groovy
@@ -1,0 +1,3 @@
+@org.osgi.annotation.bundle.Export
+@org.osgi.annotation.versioning.Version("20.0.0")
+package be.orbinson.aem.groovy.console.streaming;
