+++ bundle/src/main/groovy/be/orbinson/aem/groovy/console/servlets/ScriptPostServlet.groovy
@@ -2,6 +2,7 @@ package be.orbinson.aem.groovy.console.servlets
 
 import be.orbinson.aem.groovy.console.GroovyConsoleService
 import be.orbinson.aem.groovy.console.api.context.ScriptContext
+import be.orbinson.aem.groovy.console.api.context.impl.AsyncScriptContext
 import be.orbinson.aem.groovy.console.api.context.impl.RequestScriptContext
 import be.orbinson.aem.groovy.console.configuration.ConfigurationService
 import be.orbinson.aem.groovy.console.streaming.ExecutionRegistry
@@ -49,7 +50,18 @@ class ScriptPostServlet extends AbstractJsonResponseServlet {
                 def script = Objects.requireNonNull(getScript(request, request.getParameter(SCRIPT_PATH)),
                         "Script cannot be empty.")
                 def resourceResolver = request.resourceResolver.clone(null)
-                def executionId = executionRegistry.start(resourceResolver, script, request.getParameter(DATA))
+                def outputStream = new ByteArrayOutputStream()
+
+                def scriptContext = new AsyncScriptContext(
+                        resourceResolver: resourceResolver,
+                        outputStream: outputStream,
+                        printStream: new PrintStream(outputStream, true, StandardCharsets.UTF_8.name()),
+                        script: script,
+                        data: request.getParameter(DATA),
+                        userId: resourceResolver.userID
+                )
+
+                def executionId = executionRegistry.start(scriptContext)
 
                 writeJsonResponse(response, [executionId: executionId])
                 return
