--- it.tests/src/main/features/groovyconsole.json
@@ -1,7 +1,0 @@
-{
-  "content-packages:ARTIFACTS|required":[
-    {
-      "id": "be.orbinson.aem:aem-groovy-console-all:zip:${project.version}"
-    }
-  ]
-}
